requirements.txt
aws-cdk.core>=1.134.0
aws-cdk.assertions>=1.134.0
aws_cdk.aws_apigateway
aws_cdk.aws_lambda
aws_cdk.aws_s3
aws_cdk.aws_lambda_python